add loadswf overrider
